refactor(app): extract headless-approve-delete from headless.ts
